Keep recipe generation alive across tab switches
Move the generation state and request out of the GenerateFlow route into
a RecipeGenerationProvider mounted above AppLayout. The /generate route
unmounts on tab switch, which orphaned the in-flight request and its state
updates; the provider stays mounted across tabs, so results that arrive
while on another tab are kept and shown on return.
